[sweet API][Kotlin] Pass more information about expected types to C++ (#18653)
# Why

Another step to stop using `folly` on Android.
Passes more information about expected types to C++ to make the conversion there smarter. 
Note that this PR is just a first step in that direction and that information isn't used in many cases. However, I've made a simple optimization to create classes like Integer and Float directly in C++. Before, we alwayse converted numbers to Double and then we did a second conversion. As a result, we created more objects than needed. 

# How

- Added a class representing information about the expected type
- Added simple Integer/Float optimization. 

# Test Plan

- unit tests ✅