[barcode-scanner] Fix camera not being correctly released on unmount (Android) (#18768)
# Why

As you can see in [this snack](https://snack.expo.dev/@stefan-5gpay/basic-barcodescanner-usage) and as I've mentioned in #18746 the camera currently does not get released correctly. This means the camera indicator always stays active and the camera continues to drain battery in the background.

# How

In the code the `camera` simply was never assigned in `BarCodeScannerViewFinder`. This means the code which stops the camera was never actually run correctly because it has a null-check.

I simply assigned `camera` and added an additional check that you can't start the camera while it is stopping. With those changes in place it works.

# Test Plan

I've created the [following patch](https://gist.github.com/stefan-schweiger/a302ec60f83779f4678dbc47085e8bb0) (which does the same thing as my PR) and created a devclient with the code from the snack. When I now scan a QR code and the camera gets unmounted it works correctly

# Checklist

- [x] Documentation is up to date to reflect these changes (eg: https://docs.expo.dev and README.md).
- [x] Conforms with the [Documentation Writing Style Guide](https://github.com/expo/expo/blob/main/guides/Expo%20Documentation%20Writing%20Style%20Guide.md)
- [x] This diff will work correctly for `expo prebuild` & EAS Build (eg: updated a module plugin).