[modules][Android] Hide `Either` type behind opt-in flag (#19168)
# Why

Hides the `Either` type behind the opt-in flag. 
To indicate that some parts of Sweet API are experimental and may change, I would like to use the opt-in mechanism introduced by Kotlin. In that way, we inform users that some functions may be dangerous to use.

# How

You can read more about experimental flags here: https://kotlinlang.org/docs/opt-in-requirements.html.

# Test Plan

- not required ✅